pyproject.toml
[project]
name = "examexam"
version = "0.1.2"
description = "Let LLM generate tests, validate test, and then you take them. Useful for certs."
authors = [{ name = "Matthew Martin", email = "matthewdeanmartin@gmail.com" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "markdown",
    "scipy",
    "anthropic>=0.19.2",
    "boto3>=1.34.44",
    "colorlog>=6.8.2",
    "google-generativeai>=0.8.4",
    "mistralai>=1.6.0",
    "openai>0.28",
    "python-dotenv>=1.0.1",
    "retry>=0.9.2",
    "rtoml>=0.1.0",
    "textual>=1.0.0",
    "toml>=0.10.2",
]

[project.scripts]
examexam = "examexam.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
#
# contract
# griffe = ">=0.38.1"

[dependency-groups]
dev = [
    "cli-tool-audit",
    "git2md; python_version >= '3.10'",
    "pyclean; python_version >= '3.12'",
    "strip-docs>=1.0; python_version >= '3.12'",
    "gha-update; python_version >= '3.12'",
    "mkdocstrings[python]",
    "mkdocs; python_version >= '3.12'",
    "mdformat",
    # plugin finder
    "packaging; python_version >= '3.8'",
    # mpy
    "mypy; python_version >= '3.8'",
    "types-toml; python_version >= '3.8'",
    # reports

    # build
    "vermin; python_version >= '3.8'",
    "metametameta>=0.1.3; python_version >= '3.9'",
    "hatchling; python_version >= '3.8'",
    "ruff>=0.12.0; python_version >= '3.8'",
    "pylint; python_version >= '3.8'",
    "bandit; python_version >= '3.8'",
    "pre-commit; python_version >= '3.8'",
    # testing tools
    "pytest; python_version >= '3.8'",
    "pytest-cov; python_version >= '3.8'",
    "pytest-xdist>=3.5.0; python_version >= '3.8'",
    "pytest-randomly>=3.15.0; python_version >= '3.8'",
    "pytest-sugar>=0.9.7; python_version >= '3.8'",
    "pytest-mock; python_version >= '3.8'",
    "pytest-unused-fixtures; python_version >= '3.10'",
    "hypothesis[cli]; python_version >= '3.8'",
    "detect-test-pollution",
    "pytest-timeout>=2.4.0",
    # docs
    "interrogate>=1.5.0; python_version >= '3.8'",
    "pydoctest==0.2.1; python_version >= '3.8'",
    "pdoc3>=0.5.0; python_version >= '3.8'",
    "mdformat>=0.5.0; python_version >= '3.8'",
    "linkcheckmd>=1.4.0; python_version >= '3.8'",
    "codespell>=2.2.6; python_version >= '3.8'",
    "pyenchant>=3.2.2; python_version >= '3.8'",
]

[tool.black]
line-length = 120
target-version = ['py39']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = [
    "test",
    "tests"
]
junit_family = "xunit1"
norecursedirs = ["vendor", "scripts"]
# don't know how to do this in toml
#addopts = "--strict-markers"
#markers =
#	slow: marks tests as slow (deselect with '-m "not slow"')
#	fast: marks tests as fast (deselect with '-m "not fast"')

[tool.isort]
default_section = "THIRDPARTY"
force_grid_wrap = 0
include_trailing_comma = true
known_first_party = ["dedlin"]
line_length = 120
multi_line_output = 3
use_parentheses = true

[tool.ruff]
line-length = 1000

# Enable Pyflakes `E` and `F` codes by default.
lint.select = ["E", "F",
    "D", # docs
    "UP", # pyupgrade
    "B", # bugbear
]
lint.ignore = [
    "D402",
    "D301", # messes with doctests
    "E722",
    # only track structural problems with docstrings, not whitespace, nor missing
    "D200", "D212", "D415", "D411", "D205", "D102", "D103", "D100", "D107", "D101", "D104", "D105",
    "UP007", "UP038", # Not sure if | is well supported in lowever versions of python
]
lint.per-file-ignores = { }

# Allow unused variables when underscore-prefixed.
lint.dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Exclude a variety of commonly ignored directories.
exclude = [
    "dead_code",
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Assume Python 3.10.
target-version = "py311"


[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.scriv]
version = "literal: pyproject.toml: tool.poetry.version"
format = "md"
main_branches = "master, main, develop, dev"


[tool.coverage.report]
# Regexes for lines to exclude from consideration
exclude_also = [
    # Don't complain about missing debug-only code:
    "def __repr__",
    "if self\\.debug",
    # Don't complain if tests don't hit defensive assertion code:
    "raise AssertionError",
    "raise NotImplementedError",
    # Don't complain if non-runnable code isn't run:
    "if 0:",
    "if __name__ == .__main__.:",
    # Don't complain about abstract methods, they aren't run:
    "@(abc\\.)?abstractmethod",
]

[tool.cli-tools]
python = {name = "python", version = ">=3.11.1", schema = "semver", tags = ["build"]}
# pipx = {name = "pipx", version = ">=1.0.0", version_switch = "--version", schema = "semver"}
mypy = {name = "mypy", version = ">=1.0.0", schema = "semver", tags = ["build", "work"], install_command = "pipx install mypy", install_docs = "https://mypy.readthedocs.io/en/stable/getting_started.html"}
pylint = {name = "pylint", version = ">=1.0.0", schema = "semver", tags = ["build", "work", "user"]}
black = {name = "black", version = ">=1.0.0", schema = "semver"}
ruff = {name = "ruff", version = ">=0.6.2", schema = "semver"}
# vulture = {name = "vulture", version = "*", schema = "semver"}
# poetry = {name = "poetry", version = ">=1.5.1", schema = "semver"}



[tool.strict-build-script]
minimum_test_coverage = 35
